Move GUI IPC code to GUI lib
