+++ src/lib/gui/ElevateMode.h
