+++ src/lib/gui/IpcReader.cpp
