+++ src/lib/gui/IpcReader.h
