+++ src/lib/gui/QIpcClient.cpp
